requirements.txt
torch>=1.9.0
torchvision
transformers>=4.20.0
diffusers
datasets
Pillow>=8.3.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.25.0
gradio>=3.40.0
opencv-python>=4.5.0
librosa>=0.9.0
requests>=2.26.0
jupyter>=1.0.0
ipykernel>=6.0.0
scikit-learn>=1.0.0
